module_requirement.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg
django-cors-headers==4.3.1
